rename TimedString to CachedBytes
